THAR BE INFINITE MAPS NOW
TODO: DeSpawn entities out of range, spawn entities in range, save player spawn point

+++ SMP/Inventory/Inventory.cs
@@ -17,7 +17,7 @@ namespace SMP
 				return;
 			items[slot].item = item;
 		}
-		public void Add(Items item, int count, int slot)
+		public void Add(Items item, byte count, int slot)
 		{	
 			Item temp = new Item(item);
 			temp.count = count;
